Added Graph Evaluators
